+++ package.js
@@ -1,6 +1,6 @@
 Package.describe({
   name: 'jam:easy-schema',
-  version: '1.4.0',
+  version: '1.5.0',
   summary: 'An easy way to add schema validation for Meteor apps',
   git: 'https://github.com/jamauro/easy-schema',
   documentation: 'README.md'
@@ -16,17 +16,20 @@ Package.onUse(function(api) {
   api.use('check');
   api.use('ecmascript');
   api.use('mongo-id');
+  api.use('ddp-client');
   api.use('mdg:validation-error@0.5.1');
   api.use('zodern:types@1.0.13');
-  api.mainModule('easy-schema-client.js', 'client');
-  api.mainModule('easy-schema-server.js', 'server');
+  api.mainModule('client.js', 'client');
+  api.mainModule('server.js', 'server');
 });
 
 Package.onTest(function(api) {
   api.use('ecmascript');
   api.use('tinytest');
-  api.use('jam:easy-schema');
   api.use('mongo');
   api.use('mongo-decimal');
-  api.mainModule('easy-schema-tests.js');
+  api.use('ddp-client');
+  api.use('accounts-base');
+  api.use('jam:easy-schema');
+  api.mainModule('tests.js');
 });